Fix avatar Chooser
